LPD-18896 Remove configuration wrapper design, component should directly rely on its own configuration.
